Moving files to a place where I can write tests against it
