+++ src/GitHub.Api/GitHub.Api.csproj
@@ -221,6 +221,10 @@
     <Compile Include="Extensions\UriExtensions.cs" />
     <Compile Include="Platform\Platform.cs" />
     <Compile Include="Git\GitCredentialManager.cs" />
+    <Compile Include="UI\CommitState.cs" />
+    <Compile Include="UI\FileTreeNode.cs" />
+    <Compile Include="UI\GitCommitTarget.cs" />
+    <Compile Include="UI\TreeBuilder.cs" />
   </ItemGroup>
   <Choose>
     <When Condition="$(Buildtype) == 'Internal'">
