+++ src/GitHub.Api/UI/CommitState.cs
@@ -1,0 +1,9 @@
+namespace GitHub.Unity
+{
+    enum CommitState
+    {
+        None,
+        Some,
+        All
+    }
+}
